began work on batlogger abstract class
