+++ testADC.py
